+++ apps/backend/src/modules/users/user.service.ts
@@ -15,14 +15,15 @@ import { UserError } from "./libs/exceptions/exceptions.js";
 import {
 	type AuthenticatedUserPatchRequestDto,
 	type AuthenticatedUserPatchResponseDto,
+	type UserDetailsWithPassword,
 	type UserGetByIdItemResponseDto,
-	type UserPasswordDetails,
 	type UserPublicProfileResponseDto,
 	type UserSignUpRequestDto,
 } from "./libs/types/types.js";
 
 class UserService implements Service {
 	private groupService: GroupService;
+
 	private userRepository: UserRepository;
 
 	public constructor(
@@ -40,7 +41,7 @@ class UserService implements Service {
 
 		if (existingUser) {
 			throw new UserError({
-				message: UserExceptionMessage.USER_ALLREADY_EXISTS,
+				message: UserExceptionMessage.ALREADY_EXISTS,
 				status: HTTPCode.CONFLICT,
 			});
 		}
@@ -90,7 +91,7 @@ class UserService implements Service {
 
 	public async findPasswordDetails(
 		email: string,
-	): Promise<null | UserPasswordDetails> {
+	): Promise<null | UserDetailsWithPassword> {
 		return await this.userRepository.findPasswordDetails(email);
 	}
 
@@ -102,7 +103,7 @@ class UserService implements Service {
 
 		if (!user) {
 			throw new UserError({
-				message: UserExceptionMessage.USER_NOT_FOUND,
+				message: UserExceptionMessage.NOT_FOUND,
 				status: HTTPCode.NOT_FOUND,
 			});
 		}
@@ -118,7 +119,7 @@ class UserService implements Service {
 
 		if (!item) {
 			throw new UserError({
-				message: UserExceptionMessage.USER_NOT_FOUND,
+				message: UserExceptionMessage.NOT_FOUND,
 				status: HTTPCode.NOT_FOUND,
 			});
 		}