+++ apps/frontend/src/libs/components/icon/libs/maps/icon-name-to-svg.map.ts
@@ -8,6 +8,7 @@ import Edit from "~/assets/images/icons/edit.svg?react";
 import Error from "~/assets/images/icons/error.svg?react";
 import EyeOff from "~/assets/images/icons/eye-off.svg?react";
 import Eye from "~/assets/images/icons/eye.svg?react";
+import Logout from "~/assets/images/icons/logout.svg?react";
 import Map from "~/assets/images/icons/map.svg?react";
 import Message from "~/assets/images/icons/message.svg?react";
 import Route from "~/assets/images/icons/route.svg?react";
@@ -30,6 +31,7 @@ const iconNameToSvg: Record<
 	error: Error,
 	eye: Eye,
 	eyeOff: EyeOff,
+	logout: Logout,
 	map: Map,
 	message: Message,
 	route: Route,