+++ apps/frontend/src/modules/explore/slices/actions.ts
@@ -2,25 +2,22 @@ import { createAsyncThunk } from "@reduxjs/toolkit";
 
 import { type Location } from "~/libs/types/types.js";
 
-import { routeApi } from "../../routes/routes.js";
+import { routesApi } from "../../routes/routes.js";
 import { type RouteGetByIdResponseDto } from "../libs/types/types.js";
 import { name as sliceName } from "./explore.slice.js";
 
 const getRoutes = createAsyncThunk<
 	RouteGetByIdResponseDto[],
-	Location,
+	Location | undefined,
 	{ rejectValue: string }
->(
-	`${sliceName}/get-routes`,
-	async ({ latitude, longitude }, { rejectWithValue }) => {
-		try {
-			const response = await routeApi.getAll({ latitude, longitude });
+>(`${sliceName}/get-routes`, async (location, { rejectWithValue }) => {
+	try {
+		const response = await routesApi.getAll(location);
 
-			return response.data;
-		} catch {
-			return rejectWithValue("Failed to fetch routes.");
-		}
-	},
-);
+		return response.data;
+	} catch {
+		return rejectWithValue("Failed to fetch routes.");
+	}
+});
 
 export { getRoutes };