+++ packages/shared/src/modules/routes/libs/types/route-get-all-item-response-dto.type.ts
@@ -9,6 +9,7 @@ type RouteGetAllItemResponseDto = {
 	name: string;
 	pois: {
 		id: number;
+		name: string;
 		visitOrder: number;
 	}[];
 };